requirements.txt
# requirements.txt — Tensor4ce Jatayu POC
# Generated: 2026-03-22 from jatayu_vnev (Python 3.13)
# All versions pinned from working venv via pip freeze.

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.135.1
uvicorn==0.42.0
uvloop==0.22.1                # high-performance async event loop
starlette==0.52.1
python-multipart==0.0.22      # for multipart form uploads

# ── Agentic Framework ─────────────────────────────────────────────────────────
crewai==1.11.0
langgraph==1.1.3
langgraph-prebuilt==1.0.8
langgraph-checkpoint==4.0.1
langgraph-sdk==0.3.12
langchain-core==1.2.20
langchain-google-genai==4.2.1
langsmith==0.7.22

# ── LLM Clients ───────────────────────────────────────────────────────────────
google-genai==1.68.0          # new SDK — use Client(api_key=...)
google-generativeai==0.8.6    # legacy SDK — kept as fallback (deprecated)
google-api-core==2.30.0
google-auth==2.49.1
litellm==1.82.6               # Vertex AI via LiteLLM (LLM_BACKEND=vertex)
openai==2.29.0                # used by litellm internally

# ── ML Models ─────────────────────────────────────────────────────────────────
xgboost==3.2.0                # credit risk model (requires libomp: brew install libomp)
scikit-learn==1.8.0           # IsolationForest (fraud), preprocessing pipelines
shap==0.51.0                  # SHAP explainability for credit + fraud agents
numba==0.64.0                 # required by shap
llvmlite==0.46.0              # required by numba
scipy==1.17.1                 # ML math utilities
joblib==1.5.3                 # model persistence

# ── Data Processing ───────────────────────────────────────────────────────────
pandas==3.0.1
numpy==2.4.3
openpyxl==3.1.5              # reads Internal_Bank_Dataset.xlsx
et_xmlfile==2.0.0             # openpyxl dependency

# ── Database ──────────────────────────────────────────────────────────────────
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11      # PostgreSQL driver (sync)
aiosqlite==0.21.0            # SQLite async (used by LangGraph checkpoint)

# ── Vector Database (ChromaDB — used by CrewAI memory) ───────────────────────
chromadb==1.1.1
lancedb==0.30.1
onnxruntime==1.24.4

# ── Pydantic & Serialization ──────────────────────────────────────────────────
pydantic==2.11.10
pydantic-settings==2.10.1
pydantic_core==2.33.2
annotated-types==0.7.0

# ── Auth & Security ───────────────────────────────────────────────────────────
PyJWT==2.12.1
bcrypt==5.0.0
cryptography==46.0.5

# ── HTTP & Networking ─────────────────────────────────────────────────────────
httpx==0.28.1
requests==2.32.5
aiohttp==3.13.3

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.1.1
PyYAML==6.0.3
rich==14.3.3
tenacity==9.1.4
tiktoken==0.12.0              # token counting for LLM calls
instructor==1.14.5            # structured LLM output (CrewAI)

# ── Document Parsing ──────────────────────────────────────────────────────────
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1

# ── MCP (Model Context Protocol) ─────────────────────────────────────────────
mcp==1.26.0

# ── OpenTelemetry (tracing — CrewAI dependency) ───────────────────────────────
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
